refactor(background.rs): make Manager generic wrt Child process
Here I'm just trying to implement some common abstracting techniques
where we reduce the amoung of code depending on side effects.

In this case, we have "waiting on a running process", which is
clearly a dependency on the external environment.

By abstracting away this specific detail, we can make the background.rs
code potentially easier to test in isolation.